Switch to build-time rendering of pages
I'm not ecstatic about going this route, but I was able to get it
working relatively quickly, and it's what I'm familiar/comfortable with.

The idea here is to use a service like Netlify that can, with minimal
setup required, basically run build.py and render the entire site as
static content.